# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API des Commentaires",
        "version": "1.0"
    },
    "paths": {
...
        "/seances": {
            "get": {
                "tags": [
                    "Séances d'Entraînement"
                ],
                "summary": "Lister toutes les séances d'entraînement",
                "operationId": "583d26a889736888fdf2021850bdc8f0",
                "responses": {
                    "200": {
                        "description": "Liste des séances d'entraînement"
                    }
                }
            },
            "post": {
                "tags": [
                    "Séances d'Entraînement"
                ],
                "summary": "Créer une nouvelle séance d'entraînement",
                "operationId": "1987f570500dcd5f523384522cf70e5c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SeanceEntrainement"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Séance d'entraînement créée avec succès"
                    }
                }
            }
        },
        "/seances/{id}": {
            "put": {
                "tags": [
                    "Séances d'Entraînement"
                ],
                "summary": "Mettre à jour une séance d'entraînement",
                "operationId": "4fe1ba31d66bdde3d143c31c2631ac11",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la séance d'entraînement",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SeanceEntrainement"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Séance d'entraînement mise à jour avec succès"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Séances d'Entraînement"
                ],
                "summary": "Supprimer une séance d'entraînement",
                "operationId": "8541390d5809cc24b2e64a0ec5540f93",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la séance d'entraînement",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Séance d'entraînement supprimée avec succès"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "SeanceEntrainement": {
                "properties": {
                    "nom": {
                        "type": "string",
                        "example": "Séance Musculation Matin"
                    },
                    "description": {
                        "type": "string",
                        "example": "Séance d'entraînement de musculation en matinée."
                    },
                    "duree": {
                        "type": "string",
                        "example": "1h"
                    },
                    "chronometre": {
                        "type": "string",
                        "example": "30 minutes"
                    },
                    "ordre": {
                        "type": "integer",
                        "example": 1
                    },
                    "programme_entrainement_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "date_mise_a_jour": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-09-21T02:48:57Z"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-09-21T02:48:57Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-09-21T02:48:57Z"
                    }
                },
                "type": "object"
            }
        }
    },
...
}